Create Passwordgenerator.py
A PASSWORD GENERATOR is a useful tool that generates strong and random passwords for users. This project aims to create a password generator application using Python, allowing users to specify the length and complexity of the password.
USER INPUT: Prompt the user to specify the desired length of the password.
GENERATE PASSWORD: use a combination of random characters to generate a password  of the specified length.
DISPLAY THE PASSWORD: print the generated password on the screen.